move application to apps folder, create templates, delete south from requirements.txt, add page pagination

+++ apps/shop/models.py
@@ -5,6 +5,6 @@ class Product(models.Model):
     name = models.CharField(max_length=250)
     slug = models.CharField(max_length=250)
     description = models.TextField()
-    price = models.DecimalField(max_digits=5, decimal_places=2)
+    price = models.IntegerField()
     created_at = models.DateTimeField(auto_now_add=True)
     modified_at = models.DateTimeField()
